+++ src/core/site.rs
@@ -2,20 +2,20 @@
 #[derive(Debug, Clone)]
 pub struct Site {
     pub id: i32,
-    pub lon: f64,
-    pub lat: f64,
-    pub elevation_km: f64,
-    pub period1: f64,
+    pub lon: f32,
+    pub lat: f32,
+    pub elevation_km: f32,
+    pub period1: f32,
     /// 剪切波速，单位为m/s
-    pub vs30: f64,
+    pub vs30: f32,
     /// 深度到剪切波速为2.5km/s的深度，单位为km
-    pub z25: Option<f64>,
+    pub z25: Option<f32>,
     /// Rupture distance (km). If None, calculated from geometry.
-    pub r_rup: Option<f64>,
+    pub r_rup: Option<f32>,
     /// Joyner-Boore distance (km). If None, calculated from geometry.
-    pub r_jb: Option<f64>,
+    pub r_jb: Option<f32>,
     /// Horizontal distance from trace (km). If None, calculated from geometry.
-    pub r_x: Option<f64>,
+    pub r_x: Option<f32>,
 }
 
 impl Site {
@@ -25,12 +25,12 @@ impl Site {
     #[allow(clippy::too_many_arguments)]
     pub fn new(
         id: i32,
-        lon: f64,
-        lat: f64,
-        elevation_km: f64,
-        period1: f64,
-        vs30: f64,
-        z25: Option<f64>,
+        lon: f32,
+        lat: f32,
+        elevation_km: f32,
+        period1: f32,
+        vs30: f32,
+        z25: Option<f32>,
         is_japan: bool,
     ) -> Self {
         let mut site = Site {
@@ -53,7 +53,7 @@ impl Site {
     }
 
     /// Set manual distances
-    pub fn with_distances(mut self, r_rup: Option<f64>, r_jb: Option<f64>, r_x: Option<f64>) -> Self {
+    pub fn with_distances(mut self, r_rup: Option<f32>, r_jb: Option<f32>, r_x: Option<f32>) -> Self {
         self.r_rup = r_rup;
         self.r_jb = r_jb;
         self.r_x = r_x;
@@ -63,7 +63,7 @@ impl Site {
     /// 获取 Z2.5 (km)
     ///
     /// 如果 z25 为 None，则根据 Campbell and Bozorgnia (2014) 的经验公式，利用 Vs30 估算 Z2.5。
-    pub fn get_z25(&self, is_japan: bool) -> f64 {
+    pub fn get_z25(&self, is_japan: bool) -> f32 {
         self.z25.unwrap_or_else(|| self.estimate_z25(is_japan))
     }
 
@@ -73,7 +73,7 @@ impl Site {
     ///
     /// # 参数
     /// * `is_japan` - 是否为日本区域。如果是，使用日本的经验公式；否则使用加州的经验公式（通常也适用于其他活跃构造区）。
-    pub fn estimate_z25(&self, is_japan: bool) -> f64 {
+    pub fn estimate_z25(&self, is_japan: bool) -> f32 {
         let ln_vs30 = self.vs30.ln();
         
         let ln_z25 = if is_japan {
